update phpunit test homecontroller
